requirements.txt
amqp==5.0.6
appdirs==1.4.3
appnope==0.1.3
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
boto3==1.18.14
botocore==1.21.14
CacheControl==0.12.6
celery==5.1.2
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.0.9
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==0.6.0
decorator==5.1.0
distlib==0.3.0
distro==1.4.0
Django==3.2.4
django-cors-headers==3.7.0
django-crontab==0.7.1
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
gunicorn==19.7.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
ipdb==0.13.9
ipython==7.29.0
jedi==0.18.0
jmespath==0.10.0
kombu==5.1.0
lockfile==0.12.2
matplotlib-inline==0.1.3
msgpack==0.6.2
numpy==1.21.2
openpyxl==3.0.7
packaging==20.3
pandas==1.3.5
parso==0.8.2
paytmchecksum==1.7.0
pep517==0.8.2
pexpect==4.8.0
phonenumbers==8.12.42
pickleshare==0.7.5
Pillow==8.4.0
progress==1.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycryptodome==3.15.0
psycopg2-binary==2.9.1
pyfcm==1.5.4
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.6
python-dateutil==2.8.2
python-dotenv==0.19.0
pytoml==0.1.21
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
retrying==1.3.3
s3transfer==0.5.0
sentry-sdk==1.9.10
six==1.14.0
sqlparse==0.4.1
toml==0.10.2
traitlets==5.1.1
twilio==7.5.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
python-i18n==0.3.9
